Add it's not splitting them between folders - again, that is good. Again - what is it then?
